requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.2.3
openpyxl==3.1.5
sqlalchemy==2.0.23
python-multipart==0.0.6
requests==2.31.0
matplotlib==3.8.2
plotly==5.17.0
kaleido==0.2.1
streamlit==1.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0 